fix(mentor): adding create agent to add as inline component
